+++ tests/utils/test_clustering.py
@@ -6,12 +6,12 @@ import pandas as pd
 
 @pytest.mark.parametrize("polis_convo_data", ["small"], indirect=True)
 def test_run_kmeans_real_data_reproducible(polis_convo_data):
-    math_data, *_, _ = polis_convo_data
+    fixture = polis_convo_data
 
-    expected_cluster_centers = [group["center"] for group in math_data["group-clusters"]]
+    expected_cluster_centers = [group["center"] for group in fixture.math_data["group-clusters"]]
     cluster_count = len(expected_cluster_centers)
 
-    projected_participants = transform_base_clusters_to_participant_coords(math_data["base-clusters"])
+    projected_participants = transform_base_clusters_to_participant_coords(fixture.math_data["base-clusters"])
     projected_participants_df = pd.DataFrame([
         {
             "participant_id": item["participant_id"],
@@ -38,13 +38,13 @@ def test_run_kmeans_real_data_reproducible(polis_convo_data):
 # This is likely due to k-smoothing holding back the k value at 3 in polismath, and we're finding the real current one.
 @pytest.mark.parametrize("polis_convo_data", ["small-with-meta"], indirect=True)
 def test_find_optimal_k_real_data(polis_convo_data):
-    math_data, *_, _ = polis_convo_data
+    fixture = polis_convo_data
     MAX_GROUP_COUNT = 5
 
     # Get centers from polismath.
-    expected_centers = [group["center"] for group in math_data["group-clusters"]]
+    expected_centers = [group["center"] for group in fixture.math_data["group-clusters"]]
 
-    projected_participants = transform_base_clusters_to_participant_coords(math_data["base-clusters"])
+    projected_participants = transform_base_clusters_to_participant_coords(fixture.math_data["base-clusters"])
     projected_participants_df = pd.DataFrame([
         {
             "participant_id": item["participant_id"],