+++ tests/utils/test_polismath.py
@@ -7,10 +7,10 @@ from reddwarf.data_loader import Loader
 # TODO: Figure out if in-conv is always equivalent. If so, remove this test.
 @pytest.mark.parametrize("polis_convo_data", ["small-no-meta", "small-with-meta", "medium-with-meta", "medium-no-meta"], indirect=True)
 def test_extract_data_from_polismath(polis_convo_data):
-    math_data, _, _, _ = polis_convo_data
-    expected = math_data["in-conv"]
+    fixture = polis_convo_data
+    expected = fixture.math_data["in-conv"]
 
-    actual_participant_ids, *_ = polismath.extract_data_from_polismath(math_data)
+    actual_participant_ids, *_ = polismath.extract_data_from_polismath(fixture.math_data)
 
     assert sorted(expected) == actual_participant_ids
 
@@ -34,12 +34,12 @@ def test_extract_data_from_polismath(polis_convo_data):
 #       For now, leaning on other conversations or CSV export mitigates the issue.
 @pytest.mark.parametrize("polis_convo_data", ["small-no-meta", "small-with-meta", "medium-with-meta", "medium-no-meta"], indirect=True)
 def test_user_vote_count_sanity_check(polis_convo_data):
-    math_data, data_path, *_, _ = polis_convo_data
-    expected_user_vote_counts = math_data["user-vote-counts"]
+    fixture = polis_convo_data
+    expected_user_vote_counts = fixture.math_data["user-vote-counts"]
     # Convert keys from string to int for later comparison.
     expected_user_vote_counts = {int(k):v for k,v in expected_user_vote_counts.items()}
 
-    loader = Loader(filepaths=[f"{data_path}/votes.json"])
+    loader = Loader(filepaths=[f"{fixture.data_dir}/votes.json"])
     real_vote_matrix = MatrixUtils.generate_raw_matrix(votes=loader.votes_data)
     actual_user_vote_counts = real_vote_matrix.count(axis="columns")